make bottom view component
